held up on arXiv
I submitted this to arXiv on January 25.  It's been "on hold" since then.  I've put in two enquiries asking what I did wrong.  They just said (once) we're volunteers, we don't provide status updates.  We'll contact you.

Now, this paper has been desk rejected by the journal we submitted it to at the same time---so arXiv is now taking longer than the journal!  This is weird.  Anyway I want to share this paper for comment, so here you go.