possible covers
The cover shows a density plot in the complex plane of the eigenvalues of all $3^{16} = 43,046,721$ ``doubly companion'' matrices of dimension $m=8$ with population $\{-1,0,1\}$. Because we were unsure of rounding error effects on the delicate structures at the left and right edges, we first computed all $2,184,139$ distinct degree $8$ characteristic polynomials and solved those accurately in Maple. In fact we need not have worried, and eigenvalue computation, which is faster, resolves the delicate structures quite well. -2.2 < x < 2.2, -2.2 < y < 2.2