relative link
I want the .html file to display when linked, not to have a link to where the .html file is being stored.  

I have changed it to <A HREF="Two-cycles.html"> blah </A>

and this didn't work, then it worked, then I changed it because it didn't look like it was working, and what I changed it to didn't work, and so I have changed it back.  The delay in updating the page does cause instability.